moved config.py
