+++ examples/evaluate_random_scatterer.py
@@ -6,7 +6,7 @@ import seaborn as sns
 from loguru import logger
 from matplotlib.colors import LogNorm
 
-from fdtdx.core.config import SimulationConfig
+from fdtdx.config import SimulationConfig
 from fdtdx.core.physics import constants
 from fdtdx.fdtd.fdtd import reversible_fdtd
 from fdtdx.objects.boundaries.initialization import BoundaryConfig, boundary_objects_from_config
