+++ examples/optimize_ceviche_corner.py
@@ -16,7 +16,7 @@ from fdtdx.constraints.module import (
     StandardToInversePermittivityRange,
     StandardToPlusOneMinusOneRange,
 )
-from fdtdx.core.config import GradientConfig, SimulationConfig
+from fdtdx.config import GradientConfig, SimulationConfig
 from fdtdx.core.physics import constants
 from fdtdx.core.physics.losses import metric_efficiency
 from fdtdx.fdtd.backward import full_backward
