+++ examples/optimize_waveguide_stitching.py
@@ -10,7 +10,7 @@ from loguru import logger
 from fdtdx.constraints.discrete import BOTTOM_Z_PADDING_CONFIG_REPEAT, BinaryMedianFilterModule, RemoveFloatingMaterial
 from fdtdx.constraints.mapping import ConstraintMapping
 from fdtdx.constraints.module import ClosestIndex, IndicesToInversePermittivities, StandardToInversePermittivityRange
-from fdtdx.core.config import GradientConfig, SimulationConfig
+from fdtdx.config import GradientConfig, SimulationConfig
 from fdtdx.core.physics import constants
 from fdtdx.core.physics.losses import metric_efficiency
 from fdtdx.fdtd.backward import full_backward
