+++ src/fdtdx/constraints/module.py
@@ -4,7 +4,7 @@ from typing import Literal, Self
 import jax
 import jax.numpy as jnp
 
-from fdtdx.core.config import SimulationConfig
+from fdtdx.config import SimulationConfig
 from fdtdx.core.jax.pytrees import ExtendedTreeClass, extended_autoinit, frozen_field, frozen_private_field
 from fdtdx.core.jax.ste import straight_through_estimator
