+++ src/fdtdx/objects/detectors/detector.py
@@ -8,7 +8,7 @@ import pytreeclass as tc
 from matplotlib.figure import Figure
 from rich.progress import Progress
 
-from fdtdx.core.config import SimulationConfig
+from fdtdx.config import SimulationConfig
 from fdtdx.core.jax.typing import SliceTuple3D
 from fdtdx.core.misc import is_on_at_time_step
 from fdtdx.core.plotting.colors import LIGHT_GREEN
